Update index.js
Updated script and added a new buildTable function which creates a blank interactive seating schematic for the user to manipulate.  Changed the getResults function to deal only with the user input and apply it to a global variable for the number of rows.  Added functionality which displays the seating array as the user selects (or deselects) seats.